dvlt 关于 getpid 的修正。这个 patch 不要白不要...
git-svn-id: http://svn.kcn.cn/repos/kbs/trunk@9374 de4889a5-8ad4-4bf3-bd87-57997c21d515

+++ kbs_bbs/libBBS/article.c
@@ -1808,7 +1808,7 @@ int add_edit_mark(char *fname, int mode, char *title, session_t* session)
 
     if ((fp = fopen(fname, "rb")) == NULL)
         return 0;
-    sprintf(outname, "tmp/%d.editpost", getpid());
+    sprintf(outname, "tmp/%d.editpost", (int)getpid());
     if ((out = fopen(outname, "w")) == NULL) {
         fclose(fp);
         return 0;
